Use u/intptr_t instead of un/signed long

+++ include/freetype/ftimage.h
@@ -32,6 +32,7 @@
 #ifndef FTIMAGE_H_
 #define FTIMAGE_H_
 
+#include <stdint.h>
 
 FT_BEGIN_HEADER
 
@@ -54,7 +55,7 @@ FT_BEGIN_HEADER
    *   the context, these can represent distances in integer font units, or
    *   16.16, or 26.6 fixed-point pixel coordinates.
    */
-  typedef signed long  FT_Pos;
+  typedef intptr_t  FT_Pos;
 
 
   /**************************************************************************
